Fix tests; fix switch
