+++ frontend/src/components/menu/UserInfo.vue
@@ -15,24 +15,14 @@
         </div>
       </button>
     </template>
-    <MainButton
-      v-if="auth.isLoggedIn"
-      class="sign-out mr-4"
-      variant="third"
-      label="Sign Out"
-      size="auto"
-      @click="signOut"
-      >{{ $t('buttons.signout') }}
-    </MainButton>
+    <UserDropdown />
   </v-menu>
 </template>
 
 <script lang="ts" setup>
-import { inject } from 'vue'
-
-import MainButton from '#components/buttons/MainButton.vue'
-import AuthService from '#src/services/AuthService'
 import { useAuthStore } from '#stores/authStore'
+
+import UserDropdown from './UserDropdown.vue'
 
 const authStore = useAuthStore()
 
@@ -42,19 +32,6 @@ const initals = name
   .map((n) => n.charAt(0))
   .join('')
 const userImage = authStore.user?.profile.picture
-
-const authService = inject<AuthService>('authService')
-
-const auth = useAuthStore()
-
-async function signOut() {
-  try {
-    await authService?.signOut()
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.log('auth error', error)
-  }
-}
 </script>
 
 <style scoped lang="scss">
